+++ cognite/client/data_classes/transformations.py
@@ -1,12 +1,13 @@
 from cognite.client.data_classes._base import *
 
+from cognite.client.data_classes.transformation_jobs import TransformationJob
+
 
 class TransformationDestination:
     """TransformationDestination has static methods to define the target resource type of a transformation
 
     Args:
         type (str): Used as data type identifier on transformation creation/retrieval.
-        schema_type (str): Used as data type identifier on schema retrieval (doesn't always coincide with type).
     """
 
     def __init__(self, type: str = None):
@@ -181,6 +182,12 @@ class Transformation(CogniteResource):
         self.owner = owner
         self.owner_is_current_user = owner_is_current_user
         self._cognite_client = cognite_client
+
+    def run(self, wait: bool = True) -> TransformationJob:
+        return self._cognite_client.transformations.run(transformation_id=self.id, wait=wait)
+
+    def run_async(self) -> Awaitable[TransformationJob]:
+        return self._cognite_client.transformations.run_async(transformation_id=self.id)
 
     @classmethod
     def _load(cls, resource: Union[Dict, str], cognite_client=None):