+++ tests/tests_integration/test_api/test_transformations.py
@@ -1,14 +1,26 @@
+import random
+import string
+
 import pytest
 
 from cognite.client import CogniteClient
-from cognite.client.data_classes import Transformation, TransformationDestination, TransformationUpdate
+from cognite.client.data_classes import Transformation, TransformationDestination, TransformationUpdate, TransformationJob, TransformationJobStatus
 
 COGNITE_CLIENT = CogniteClient()
+prefix = "".join(random.choice(string.ascii_letters) for i in range(6))
 
 
 @pytest.fixture
 def new_transformation():
-    transform = Transformation(name="any", destination=TransformationDestination.raw())
+    transform = Transformation(
+        name="any",
+        external_id=f"{prefix}-transformation",
+        destination=TransformationDestination.assets(),
+        query="select id, name from _cdf.assets limit 5000",
+        source_api_key=COGNITE_CLIENT.config.api_key,
+        destination_api_key=COGNITE_CLIENT.config.api_key,
+        ignore_null_fields=True,
+    )
     ts = COGNITE_CLIENT.transformations.create(transform)
 
     yield ts
@@ -36,8 +48,7 @@ class TestTransformationsAPI:
     def test_create(self, new_transformation):
         assert (
             new_transformation.name == "any"
-            and new_transformation.destination.type == "raw_table"
-            and new_transformation.destination.rawType == "plain_raw"
+            and new_transformation.destination.type == "assets"
             and new_transformation.id is not None
         )
 
@@ -73,3 +84,53 @@ class TestTransformationsAPI:
     def test_list(self, new_transformation):
         retrieved_transformations = COGNITE_CLIENT.transformations.list()
         assert new_transformation.id in [transformation.id for transformation in retrieved_transformations]
+
+    def test_run(self, new_transformation: Transformation):
+        job = new_transformation.run()
+
+        assert (
+            job.id is not None
+            and job.uuid is not None
+            and job.status == TransformationJobStatus.COMPLETED
+            and job.source_project == COGNITE_CLIENT.config.project
+            and job.destination_project == COGNITE_CLIENT.config.project
+            and job.destination_type == "assets"
+            and job.conflict_mode == "upsert"
+            and job.raw_query == new_transformation.query
+            and job.error is None
+            and job.ignore_null_fields
+        )
+
+    @pytest.mark.asyncio
+    async def test_run_async(self, new_transformation: Transformation):
+        job = await new_transformation.run_async()
+
+        assert (
+            job.id is not None
+            and job.uuid is not None
+            and job.status == TransformationJobStatus.COMPLETED
+            and job.source_project == COGNITE_CLIENT.config.project
+            and job.destination_project == COGNITE_CLIENT.config.project
+            and job.destination_type == "assets"
+            and job.conflict_mode == "upsert"
+            and job.raw_query == new_transformation.query
+            and job.error is None
+            and job.ignore_null_fields
+        )
+
+    @pytest.mark.asyncio
+    async def test_run_by_external_id_async(self, new_transformation: Transformation):
+        job = await COGNITE_CLIENT.transformations.run_async(transformation_external_id=f"{prefix}-transformation")
+
+        assert (
+            job.id is not None
+            and job.uuid is not None
+            and job.status == TransformationJobStatus.COMPLETED
+            and job.source_project == COGNITE_CLIENT.config.project
+            and job.destination_project == COGNITE_CLIENT.config.project
+            and job.destination_type == "assets"
+            and job.conflict_mode == "upsert"
+            and job.raw_query == new_transformation.query
+            and job.error is None
+            and job.ignore_null_fields
+        )